Add build-args to support jmalloc compilations via Docker running on macOS hosts.
If the image is built with `docker build --build-arg macos_hsot_arch=arm64` and the image detects `x86_64`, it will know its emulated and set the correct LG_VADDR variable.

If the argument is not passed then its completely ignored, so it remains compatible with existing build pipelines.

Based on the problem at https://github.com/tamatebako/tebako/pull/178